+++ syscall/samples/access/access.c
@@ -7,20 +7,18 @@
 
 int main(void)
 {
-	char dir[] = "tmpd-access";
+	char file[] = "/etc/os-release";
 
-	mkdir(dir, 0700);
-
-	if (0 != access(dir, F_OK))
+	if (access(file, F_OK) != 0)
 		fprintf(stderr, "F_OK %s\n", strerror(errno));
-	if (0 != access(dir, R_OK))
+	if (access(file, R_OK) != 0)
 		fprintf(stderr, "R_OK %s\n", strerror(errno));
-	if (0 != access(dir, W_OK))
+	if (access(file, W_OK) != 0)
 		fprintf(stderr, "W_OK %s\n", strerror(errno));
-	if (0 != access(dir, X_OK))
+	if (access(file, X_OK) != 0)
 		fprintf(stderr, "X_OK %s\n", strerror(errno));
-
-	rmdir(dir);
+	if (access(file, F_OK | R_OK) != 0)
+		fprintf(stderr, "F_OK | W_OK %s\n", strerror(errno));
 
 	return 0;
 }