+++ syscall/samples/module/delete_module.c
@@ -2,12 +2,11 @@
 #include <stdlib.h>
 #include <unistd.h>
 #include <fcntl.h>
-#include <sys/syscall.h>
 #include <errno.h>
 
 #include "mod_helpers.h"
 
-int main(void)
+int main(int argc, char *argv[])
 {
 	sys_delete_module("mymod", O_NONBLOCK);
 	return 0;