+++ syscall/samples/module/mod_helpers.c
@@ -8,6 +8,17 @@
 #include "mod_helpers.h"
 
 
+int sys_init_module(void *module_image, unsigned long len,
+		    const char *param_values)
+{
+	return syscall(SYS_init_module, module_image, len, param_values);
+}
+
+int sys_finit_module(int fd, const char *param_values, int flags)
+{
+	return syscall(SYS_finit_module, fd, param_values, flags);
+}
+
 int sys_delete_module(const char *name, unsigned int flags)
 {
 	return syscall(SYS_delete_module, name, flags);