+++ syscall/samples/module/mod_helpers.h
@@ -1,3 +1,6 @@
 #pragma once
 
+int sys_init_module(void *module_image, unsigned long len,
+		    const char *param_values);
+int sys_finit_module(int fd, const char *param_values, int flags);
 int sys_delete_module(const char *name, unsigned int flags);